backend/requirements.txt
fastapi
uvicorn
face_recognition
opencv-python-headless
faiss-cpu
psycopg2-binary
sqlalchemy
python-multipart
python-dotenv
pillow